feature(children) : enhance registration flow with CAF subscription handling

+++ app/controllers/children_controller.rb
@@ -37,6 +37,8 @@ class ChildrenController < ApplicationController
       build_child_for_form
       @child.build_children_source(source_id: children_source_params&.dig(:source_id), details: children_source_params&.dig(:details), registration_department: children_source_params&.dig(:registration_department))
       render action: :new
+    elsif current_registration_origin == 2
+      redirect_to created_child_path(child_support_id: @child.child_support.id, current_child_name: @child.first_name, sc: @child.parent1.security_code)
     elsif service.parent1_target_profile || current_registration_origin != 4
       redirect_to created_child_path(sms_url_form: service.sms_url_form, children_under_four_months: service.children_under_four_months, youngest_child_under_twenty_four_months: service.youngest_child_under_twenty_four_months)
     else
@@ -82,11 +84,15 @@ class ChildrenController < ApplicationController
       @widget = false
       @new_link = new_pmi_registration_path
     when 2
-      session.delete(:registration_origin)
-      @message = I18n.t(support_wait_time_message, typeform_url: params[:sms_url_form])
-      @again = false
-      @widget = false
-      @new_link = new_caf_registration_path
+      if ENV['CAF_SUBSCRIPTION'].present?
+        @caf_subscription_form = true
+      else
+        session.delete(:registration_origin)
+        @message = I18n.t(support_wait_time_message, typeform_url: params[:sms_url_form])
+        @again = false
+        @widget = false
+        @new_link = new_caf_registration_path
+      end
     else
       @message = I18n.t('inscription_success.with_widget')
       @again = false
